refactor(deploy-aws): clean up pipeline creation flow
